Save user orders
Succesfully captures user orders and save a max of 3 inside an array, which is later sent to the constructor function, and that is saved in local storage for later use. I´m sure there are more efficient ways to do it. However, for now, I accept this. First independant project is worth it. Time to up my game.